# contracts/openapi/gateway.openapi.yaml
openapi: 3.0.3
info:
  title: Gateway Public API
  version: "1.0.0"
servers:
  - url: "{PUBLIC_API_URI}"
    description: Public gateway base URI (injected per environment)
    variables:
      PUBLIC_API_URI:
        default: https://api.the-actual-news.example
        description: Public base URI of the gateway deployment
# Public read API — endpoints are unauthenticated by design.
security: []
paths:
  /v1/health:
    get:
      operationId: gatewayHealth
      summary: Liveness/readiness probe for the public gateway
      responses: { "200": { description: ok } }

  /v1/feed:
    get:
      operationId: getFeed
      summary: Get the public story feed for a geographic scope
      parameters:
        - in: query
          name: scope
          schema: { type: string, enum: [local, regional, national, global], default: local }
      responses:
        "200":
          description: feed
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [story_id, title, state, updated_at]
                      properties:
                        story_id: { $ref: "common.openapi.yaml#/components/schemas/Ulid" }
                        title: { type: string }
                        state: { type: string }
                        updated_at: { $ref: "common.openapi.yaml#/components/schemas/IsoTime" }

  /v1/story/{story_id}:
    get:
      operationId: getStoryBundle
      summary: Get the full reader bundle (story, claims, evidence, corrections)
      parameters:
        - in: path
          name: story_id
          required: true
          schema: { $ref: "common.openapi.yaml#/components/schemas/Ulid" }
      responses:
        "200":
          description: story bundle for reader view
          content:
            application/json:
              schema:
                type: object
                required: [story, claims, evidence_edges, corrections, bundle_hash]
                properties:
                  story: { $ref: "story.openapi.yaml#/components/schemas/StoryWithVersions" }
                  claims:
                    type: array
                    items: { $ref: "claim.openapi.yaml#/components/schemas/Claim" }
                  evidence_edges:
                    type: array
                    items:
                      type: object
                      required: [claim_id, evidence_id_hash, relation, strength]
                      properties:
                        claim_id: { $ref: "common.openapi.yaml#/components/schemas/Ulid" }
                        evidence_id_hash: { type: string }
                        relation: { type: string, enum: [supports, contradicts, context] }
                        strength: { type: number, minimum: 0, maximum: 1 }
                  corrections:
                    type: array
                    items:
                      type: object
                      required: [correction_id, claim_id, reason, created_at]
                      properties:
                        correction_id: { $ref: "common.openapi.yaml#/components/schemas/Ulid" }
                        claim_id: { $ref: "common.openapi.yaml#/components/schemas/Ulid" }
                        reason: { type: string }
                        created_at: { $ref: "common.openapi.yaml#/components/schemas/IsoTime" }
                  bundle_hash:
                    type: string
                    description: Canonical SHA-256 hash of the public story bundle.

  /v1/story/{story_id}/transparency:
    get:
      operationId: getStoryTransparency
      parameters:
        - in: path
          name: story_id
          required: true
          schema: { $ref: "common.openapi.yaml#/components/schemas/Ulid" }
      responses:
        "200":
          description: story transparency checkpoints and inclusion proofs
          content:
            application/json:
              schema:
                type: object
                required: [story_id, bundle_hash, checkpoints]
                properties:
                  story_id: { $ref: "common.openapi.yaml#/components/schemas/Ulid" }
                  bundle_hash: { type: string }
                  checkpoints:
                    type: array
                    items:
                      type: object
                      required: [checkpoint_id, merkle_root, created_at, event_id, inclusion_proof]
                      properties:
                        checkpoint_id: { $ref: "common.openapi.yaml#/components/schemas/Ulid" }
                        merkle_root: { type: string }
                        created_at: { $ref: "common.openapi.yaml#/components/schemas/IsoTime" }
                        event_id: { $ref: "common.openapi.yaml#/components/schemas/Ulid" }
                        inclusion_proof:
                          type: object
                          required: [leaf_hash, leaf_index, proof]
                          properties:
                            leaf_hash: { type: string }
                            leaf_index: { type: integer, minimum: 0 }
                            proof:
                              type: array
                              items:
                                type: object
                                required: [side, hash]
                                properties:
                                  side: { type: string, enum: [left, right] }
                                  hash: { type: string }

  /v1/transparency/checkpoints:
    post:
      operationId: createTransparencyCheckpoint
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                limit: { type: integer, minimum: 1, maximum: 1000, default: 100 }
      responses:
        "201":
          description: checkpoint created
        "200":
          description: no eligible events

  /v1/transparency/checkpoints/{checkpoint_id}:
    get:
      operationId: getTransparencyCheckpoint
      parameters:
        - in: path
          name: checkpoint_id
          required: true
          schema: { $ref: "common.openapi.yaml#/components/schemas/Ulid" }
      responses:
        "200":
          description: transparency checkpoint